use a common grammar for staterestore tests
