+++ test/staterestore/standard/staterestore_test.go
@@ -1,4 +1,4 @@
-package staterestoreopt
+package staterestore
 
 import "testing"
