+++ test/staterestore/staterestore.peg
@@ -11,7 +11,7 @@
       defer f.Close()
       in = f
     }
-    got, err := ParseReader("", in, Debug(false))
+    got, err := ParseReader("", in)
     fmt.Println(got, err)
   }
 }
